Refactor unit test infrastructure
