+++ dc-cudami-server/dc-cudami-server-webapp/pom.xml
@@ -90,6 +90,11 @@
       <scope>runtime</scope>
     </dependency>
     <dependency>
+      <groupId>org.assertj</groupId>
+      <artifactId>assertj-core</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>org.flywaydb</groupId>
       <artifactId>flyway-core</artifactId>
     </dependency>
