+++ dc-cudami-server/dc-cudami-server-webapp/src/test/java/de/digitalcollections/cudami/server/ActuatorControllerTest.java
@@ -1,11 +1,9 @@
-package de.digitalcollections.cudami.server.test.controller;
+package de.digitalcollections.cudami.server;
 
 import static org.assertj.core.api.BDDAssertions.then;
 
-import de.digitalcollections.cudami.server.test.TestApplication;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.actuate.autoconfigure.web.server.LocalManagementPort;
@@ -13,15 +11,9 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 /** Basic integration tests for service application. */
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(
-    classes = {TestApplication.class},
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestPropertySource(properties = {"spring.profiles.active=local", "management.server.port=0"})
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ActuatorControllerTest {
 
   @Value("${management.endpoints.web.base-path}")
