+++ dc-cudami-server/dc-cudami-server-webapp/src/test/java/de/digitalcollections/cudami/server/config/SpringConfigBackendForTest.java
@@ -1,16 +1,16 @@
-package de.digitalcollections.cudami.server.test.config;
+package de.digitalcollections.cudami.server.config;
 
 import javax.sql.DataSource;
 import org.mockito.Mockito;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 
-@Configuration
+@TestConfiguration
 public class SpringConfigBackendForTest {
 
-  @Primary
   @Bean
+  @Primary
   public DataSource dataSource() {
     return Mockito.mock(DataSource.class);
   }
